profile image upload done, user page dynamic now

+++ lib/models/babysitter_model.dart
@@ -3,7 +3,7 @@ import 'package:babysitter_booking_app/models/user_model.dart';
 
 class Babysitter extends UserModel {
   String id, email, password, name, street, county;
-  double wage;
+  int wage;
   List<Job> jobs;
 
   Babysitter() : super() {
